Allow customizable line breaks
By default, new lines are used rather than pipes. This data is stored in a new attribute to the `script` tag: `linebreak`. To preserve the original pipe character being used, if the attribute is not present, the level will be loaded in using pipes. It will, however, be saved using new lines.

TODO: New lines should appear as `"\n"` in the XML file. Right now they appear as an actual new line.